+++ src/actinia_core/rest/base/renderer_base.py
